jobEnd

+++ taskfactory-biz/src/main/java/com/mk/taskfactory/biz/impl/RoomTypeInfoServiceImpl.java
@@ -3,7 +3,6 @@ package com.mk.taskfactory.biz.impl;
 import com.mk.taskfactory.api.RoomTypeInfoService;
 import com.mk.taskfactory.api.dtos.TRoomTypeInfoDto;
 import com.mk.taskfactory.biz.mapper.RoomTypeInfoMapper;
-import com.mk.taskfactory.model.TRoomTypeInfo;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,7 +14,10 @@ public class RoomTypeInfoServiceImpl implements RoomTypeInfoService {
     private RoomTypeInfoMapper roomTypeInfoMapper;
 
     @Override
-    public void deleteByRoomType(int roomTypeId) {
+    public void deleteByRoomType(Integer roomTypeId) {
+        if(roomTypeId==null){
+            return;
+        }
         this.roomTypeInfoMapper.deleteByRoomType(roomTypeId);
     }
     public TRoomTypeInfoDto findByRoomTypeId(Integer roomTypeId){
